% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv_3yrotation.R
\name{npv_3yrotation}
\alias{npv_3yrotation}
\title{Net present value of drainage for 3-year crop rotation}
\usage{
npv_3yrotation(np1, np2, np3, wl, dc, r, t)
}
\arguments{
\item{np1}{It is the difference between farm revenue and cost of production for crop 1.Unit is $/ha.}

\item{np2}{It is the difference between farm revenue and cost of production for crop 2.Unit is $/ha.}

\item{np3}{It is the difference between farm revenue and cost of production for crop 3.Unit is $/ha.}

\item{wl}{Wetland size in ha.}

\item{dc}{Drainage cost in $.}

\item{r}{Discount rate Percentage.}

\item{t}{Crop production planning horizon.}
}
\value{
Single value or dataframe of net present value of converted wetland(s).
}
\description{
Net present value of drainage for 3-year crop rotation
}
\examples{
npv_3yrotation(np1=10, np2=20, np3=30,wl=3, dc=500, r=0.87, t=50)
}
