% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv_5yrotation.R
\name{npv_5yrotation}
\alias{npv_5yrotation}
\title{Net present value of drainage for 5-year crop rotation}
\usage{
npv_5yrotation(np1, np2, np3, np4, np5, wl, dc, r, t)
}
\arguments{
\item{np1}{It is the difference between farm revenue and cost of production for crop 1. Unit is $/ha.}

\item{np2}{It is the difference between farm revenue and cost of production for crop 2. Unit is $/ha.}

\item{np3}{It is the difference between farm revenue and cost of production for crop 3. Unit is $/ha.}

\item{np4}{It is the difference between farm revenue and cost of production for crop 4. Unit is $/ha.}

\item{np5}{It is the difference between farm revenue and cost of production for crop 5. Unit is $/ha.}

\item{wl}{Wetland size in ha.}

\item{dc}{Drainage cost in $.}

\item{r}{Discount rate in Percentage.}

\item{t}{Crop production planning horizon.}
}
\value{
The output can be a dataframe or single vale of net present value of converted wetland(s).
}
\description{
Net present value of drainage for 5-year crop rotation
}
\examples{
npv_5yrotation(np1=10, np2=20, np3=30, np4=40, np5=50, wl=3, dc=500, r=0.87, t=50)
}
