% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_rotation.R
\name{prop_rotation}
\alias{prop_rotation}
\title{Net present value of drainage for proportional crop rotation}
\usage{
prop_rotation(
  y1np1,
  y1np2,
  y1np3,
  y1np4,
  y1np5,
  y2np1,
  y2np2,
  y2np3,
  y2np4,
  y2np5,
  y3np1,
  y3np2,
  y3np3,
  y3np4,
  y3np5,
  y4np1,
  y4np2,
  y4np3,
  y4np4,
  y4np5,
  y5np1,
  y5np2,
  y5np3,
  y5np4,
  y5np5,
  y1pwlnp1,
  y1pwlnp2,
  y1pwlnp3,
  y1pwlnp4,
  y1pwlnp5,
  y2pwlnp1,
  y2pwlnp2,
  y2pwlnp3,
  y2pwlnp4,
  y2pwlnp5,
  y3pwlnp1,
  y3pwlnp2,
  y3pwlnp3,
  y3pwlnp4,
  y3pwlnp5,
  y4pwlnp1,
  y4pwlnp2,
  y4pwlnp3,
  y4pwlnp4,
  y4pwlnp5,
  y5pwlnp1,
  y5pwlnp2,
  y5pwlnp3,
  y5pwlnp4,
  y5pwlnp5,
  wl,
  dc,
  r,
  t
)
}
\arguments{
\item{y1np1}{Net profit for crop 1 in year 1. Unit is $/ha.}

\item{y1np2}{Net profit for crop 2 in year 1. Unit is $/ha.}

\item{y1np3}{Net profit for crop 3 in year 1. Unit is $/ha.}

\item{y1np4}{Net profit for crop 4 in year 1. Unit is $/ha.}

\item{y1np5}{Net profit for crop 5 in year 1. Unit is $/ha.}

\item{y2np1}{Net profit for crop 1 in year 2. Unit is $/ha.}

\item{y2np2}{Net profit for crop 2 in year 2. Unit is $/ha.}

\item{y2np3}{Net profit for crop 3 in year 2. Unit is $/ha.}

\item{y2np4}{Net profit for crop 4 in year 2. Unit is $/ha.}

\item{y2np5}{Net profit for crop 5 in year 2. Unit is $/ha.}

\item{y3np1}{Net profit for crop 1 in year 3. Unit is $/ha.}

\item{y3np2}{Net profit for crop 2 in year 3. Unit is $/ha.}

\item{y3np3}{Net profit for crop 3 in year 3. Unit is $/ha.}

\item{y3np4}{Net profit for crop 4 in year 3. Unit is $/ha.}

\item{y3np5}{Net profit for crop 5 in year 3. Unit is $/ha.}

\item{y4np1}{Net profit for crop 1 in year 4. Unit is $/ha.}

\item{y4np2}{Net profit for crop 2 in year 4. Unit is $/ha.}

\item{y4np3}{Net profit for crop 3 in year 4. Unit is $/ha.}

\item{y4np4}{Net profit for crop 4 in year 4. Unit is $/ha.}

\item{y4np5}{Net profit for crop 5 in year 4. Unit is $/ha.}

\item{y5np1}{Net profit for crop 1 in year 5. Unit is $/ha.}

\item{y5np2}{Net profit for crop 2 in year 5. Unit is $/ha.}

\item{y5np3}{Net profit for crop 3 in year 5. Unit is $/ha.}

\item{y5np4}{Net profit for crop 4 in year 5. Unit is $/ha.}

\item{y5np5}{Net profit for crop 5 in year 5. Unit is $/ha.}

\item{y1pwlnp1}{Proportion of wetland area devoted to crop 1 in year 1. It should be between 0 and 1 inclusive.}

\item{y1pwlnp2}{Proportion of wetland area devoted to crop 2 in year 1. It should be between 0 and 1 inclusive.}

\item{y1pwlnp3}{Proportion of wetland area devoted to crop 3 in year 1. It should be between 0 and 1 inclusive.}

\item{y1pwlnp4}{Proportion of wetland area devoted to crop 4 in year 1. It should be between 0 and 1 inclusive.}

\item{y1pwlnp5}{Proportion of wetland area devoted to crop 5 in year 1. It should be between 0 and 1 inclusive.}

\item{y2pwlnp1}{Proportion of wetland area devoted to crop 1 in year 2. It should be between 0 and 1 inclusive.}

\item{y2pwlnp2}{Proportion of wetland area devoted to crop 2 in year 2. It should be between 0 and 1 inclusive.}

\item{y2pwlnp3}{Proportion of wetland area devoted to crop 3 in year 2. It should be between 0 and 1 inclusive.}

\item{y2pwlnp4}{Proportion of wetland area devoted to crop 4 in year 2. It should be between 0 and 1 inclusive.}

\item{y2pwlnp5}{Proportion of wetland area devoted to crop 5 in year 2. It should be between 0 and 1 inclusive.}

\item{y3pwlnp1}{Proportion of wetland area devoted to crop 1 in year 3. It should be between 0 and 1 inclusive.}

\item{y3pwlnp2}{Proportion of wetland area devoted to crop 2 in year 3. It should be between 0 and 1 inclusive.}

\item{y3pwlnp3}{Proportion of wetland area devoted to crop 3 in year 3. It should be between 0 and 1 inclusive.}

\item{y3pwlnp4}{Proportion of wetland area devoted to crop 4 in year 3. It should be between 0 and 1 inclusive.}

\item{y3pwlnp5}{Proportion of wetland area devoted to crop 5 in year 3. It should be between 0 and 1 inclusive.}

\item{y4pwlnp1}{Proportion of wetland area devoted to crop 1 in year 4. It should be between 0 and 1 inclusive.}

\item{y4pwlnp2}{Proportion of wetland area devoted to crop 2 in year 4. It should be between 0 and 1 inclusive.}

\item{y4pwlnp3}{Proportion of wetland area devoted to crop 3 in year 4. It should be between 0 and 1 inclusive.}

\item{y4pwlnp4}{Proportion of wetland area devoted to crop 4 in year 4. It should be between 0 and 1 inclusive.}

\item{y4pwlnp5}{Proportion of wetland area devoted to crop 5 in year 4. It should be between 0 and 1 inclusive.}

\item{y5pwlnp1}{Proportion of wetland area devoted to crop 1 in year 5. It should be between 0 and 1 inclusive.}

\item{y5pwlnp2}{Proportion of wetland area devoted to crop 2 in year 5. It should be between 0 and 1 inclusive.}

\item{y5pwlnp3}{Proportion of wetland area devoted to crop 3 in year 5. It should be between 0 and 1 inclusive.}

\item{y5pwlnp4}{Proportion of wetland area devoted to crop 4 in year 5. It should be between 0 and 1 inclusive.}

\item{y5pwlnp5}{Proportion of wetland area devoted to crop 5 in year 5. It should be between 0 and 1 inclusive.}

\item{wl}{Wetland area (ha).}

\item{dc}{Drainage Cost ($/ha).}

\item{r}{Discount rate (percentage).}

\item{t}{Planning horizon (years).}
}
\value{
The output can be a dataframe or single vale of net present value of converted wetland(s).
}
\description{
Net present value of drainage for proportional crop rotation
}
\examples{
prop_rotation <- function(y1np1 =10, y1np2=30, y1np3=40, y1np4=50, y1np5=60,
y1pwlnp1=0.2, y1pwlnp2=0.2, y1pwlnp3=0.2, y1pwlnp4=0.2, y1pwlnp5=0.2,
y2np1=70, y2np2 =80, y2np3=0, y2np4=0, y2np5=0,
y2pwlnp1=0.6, y2pwlnp2=0.4, y2pwlnp3=0, y2pwlnp4=0, y2pwlnp5=0,
y3np1=90, y3np2=0, y3np3=0, y3np4=0, y3np5,
y3pwlnp1=1, y3pwlnp2=0, y3pwlnp3=0, y3pwlnp4=0, y3pwlnp5=0,
y4np1=90, y4np2=0, y4np3=0, y4np4=0, y4np5=0,
y4pwlnp1=1, y4pwlnp2=0, y4pwlnp3=0, y4pwlnp4=0, y4pwlnp5=0,
y5np1=30, y5np2=30, y5np3=80, y5np4=0, y5np5=0,
y5pwlnp1=0.5, y5pwlnp2=0.3, y5pwlnp3=0.2, y5pwlnp4=0, y5pwlnp5=0,
wl=10, dc=400, r=0.8, t=50)
}
