% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlandscape.R
\name{wetlandscape}
\alias{wetlandscape}
\title{Wetland landscape generator}
\usage{
wetlandscape(
  qsect,
  t1ws_min,
  t1wls_max,
  t3ws_min,
  t3wls_max,
  t6ws_min,
  t6wls_max,
  t12ws_min,
  t12wls_max,
  t24ws_min,
  t24wls_max,
  t1pr_min,
  t3pr_min,
  t6pr_min,
  t12pr_min,
  t24pr_min,
  dc_min,
  dc_max
)
}
\arguments{
\item{qsect}{Integer. Full name is quarter-section and it is the number of
quarter-section (160-acre) farmlands in the sub-basin}

\item{t1ws_min}{Integer. It is the minimum wetland size (ha) in the tier 1 group of wetlands in the sub-basin.}

\item{t1wls_max}{Integer. It is the maximum wetland size (ha) in the tier 1 group of wetlands in the sub-basin.}

\item{t3ws_min}{Integer. It is the minimum wetland size (ha) in the tier 3 group of wetlands in the sub-basin.}

\item{t3wls_max}{Integer. It is the maximum wetland size (ha) in the tier 3 group of wetlands in the sub-basin.}

\item{t6ws_min}{Integer. It is the minimum wetland size (ha) in the tier 6 group of wetlands in the sub-basin.}

\item{t6wls_max}{Integer. It is the maximum wetland size (ha) in the tier 6 group of wetlands in the sub-basin.}

\item{t12ws_min}{Integer. It is the minimum wetland size (ha) in the tier 12 group of wetlands in the sub-basin.}

\item{t12wls_max}{Integer. It is the maximum wetland size (ha) in the tier 12 group of wetlands in the sub-basin.}

\item{t24ws_min}{Integer. It is the minimum wetland size (ha) in the tier 24 group of wetlands in the sub-basin.}

\item{t24wls_max}{Integer. It is the maximum wetland size (ha) in the tier 24 group of wetlands in the sub-basin.}

\item{t1pr_min}{Integer. It is the minimum probability of harvesting in the tier 1 group of wetlands in the sub-basin.
It is between 0 and 1.}

\item{t3pr_min}{Integer. It is the minimum probability of harvesting in the tier 3 group of wetlands in the sub-basin.
It is between 0 and 1.}

\item{t6pr_min}{Integer. It is the minimum probability of harvesting in the tier 6 group of wetlands in the sub-basin.
It is between 0 and 1.}

\item{t12pr_min}{Integer. It is the minimum probability of harvesting in the tier 12 group of wetlands in the sub-basin.
It is between 0 and 1.}

\item{t24pr_min}{Integer. It is the minimum probability of harvesting in the tier 24 group of wetlands in the sub-basin.
It is between 0 and 1.}

\item{dc_min}{Integer. It is the minimum drainage cost of wetlands, $/ha.}

\item{dc_max}{Integer. It is the maximum drainage cost of wetlands, $/ha.}
}
\value{
The output is a dataframe of wetland characteristics in the sub-basin.
}
\description{
Wetland landscape generator
}
\examples{
wetlandscape(qsect = 14000, t1ws_min = 0.1, t1wls_max = 1,
t3ws_min = 0.5, t3wls_max = 2, t6ws_min = 3, t6wls_max = 5, t12ws_min=3, t12wls_max=4,
t24ws_min=2, t24wls_max=6,t1pr_min=0.6, t3pr_min=0.5, t6pr_min=0.6, t12pr_min=0.7, t24pr_min = .9, dc_min = 500, dc_max = 1000)
}
